adding an alertdialog before deleting recipe

+++ app/src/main/java/be/aryssimon/cookbook/MainActivity.kt
@@ -1,6 +1,7 @@
 package be.aryssimon.cookbook
 
 import android.annotation.SuppressLint
+import android.content.DialogInterface
 import android.content.Intent
 import android.content.res.Resources
 import android.os.Bundle
@@ -12,10 +13,12 @@ import android.view.View
 import android.widget.Button
 import android.widget.TextView
 import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.room.Room
 import com.google.android.material.appbar.CollapsingToolbarLayout
 import com.google.android.material.floatingactionbutton.FloatingActionButton
+import com.google.android.material.snackbar.Snackbar
 import java.util.*
 
 
@@ -68,7 +71,15 @@ class MainActivity : AppCompatActivity() {
             }
             R.id.action_delete -> {
                 if (recipeList.isNotEmpty()) {
-                    deleteRecipe()
+                    AlertDialog.Builder(this@MainActivity)
+                        .setTitle(getString(R.string.delete_recipe_title))
+                        .setMessage(getString(R.string.delete_recipe_message))
+                        .setPositiveButton(getString(R.string.confirm_delete),
+                            DialogInterface.OnClickListener { dialog, which ->
+                                deleteRecipe()
+                            })
+                        .setNegativeButton(getString(R.string.cancel_delete), null)
+                        .show()
                 } else {
                     Toast.makeText(this@MainActivity, R.string.nothing_to_delete, Toast.LENGTH_SHORT).show()
                 }
